Correctly draw chunk borders
When Show Player is enabled, chunk borders were being rendered relative
to the player instead of the camera. This meant that when you flew into
another chunk, the "highlighted" chunk was still the one with the
player in it.

This might be better fixed by replacing the Show Player override in
`CameraMixin` with something more targeted & specific. Overriding an
override feels a little hacky...