"Initial Perspective" config option
Add an option to select a perspective to start FreeCam in.

- "Inside" this option retains the current behavior.
- "First Person" this option shifts the camera 0.4 blocks forward,
  landing just in front of the player's eyes.
- "Third Person" this option replicates F5 mode, moving the camera 4
  blocks behind the player.
- "Mirror" this option replicates mirrored F5 mode, moving the camera 4
  blocks in front of the player and inverting the camera.

Default is set to "Inside" to retain existing behaviour.

Also added an "Always Check Collision" option. If enabled (or No Clip
is disabled), only move the camera as far as possible before either
colliding with blocks or reaching the default distance.

Fixes #16