Use the "swimming" pose in freecam
The "swimming" pose is normally used when the player is crawling through a one-block-high space. Or when the player activates the swimming hotkey in water.

The swimming pose has the smallest player hitbox available in survival, so when noClip is disabled the swimming pose is required to get through one-block-high gaps and tunnels.

Made it so that when toggling freecam, the pose is automatically set to swimming and the freecam eye height is matched to the player's seamlessly. You cannot tell your pose has changed until you try to fit through a small gap.

The new CameraMixin skips an animation that transitions changes in pose gradually. The mixin only applies when the camera entity is being toggled to or from an instance of FreeCamera.