Add changelog management plugin
Gradle Changelog Plugin[1] by JetBrains.

Can be used to "patch" the changelog; marking the current unreleased changes as belonging to the current `mod_version` set in `gradle.properties`.

Can also be used to "get" the changelog relating to a specific previous version, which will be useful when publishing new releases.

[1]: https://github.com/JetBrains/gradle-changelog-plugin